fix: register page.tsx to be a server component
